# apps/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Lelang Ikan API",
        "description": "API untuk aplikasi e-lelang ikan realtime",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Local API Base"
        }
    ],
    "paths": {
...
        "/orders/{id}/dispute": {
            "post": {
                "tags": [
                    "Disputes"
                ],
                "summary": "Open dispute for an order (buyer/seller)",
                "operationId": "7fc43615d679e7d090372da66ae6ca27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reason"
                                ],
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Barang tidak sesuai"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Berat tidak sesuai deskripsi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/disputes": {
            "get": {
                "tags": [
                    "Disputes"
                ],
                "summary": "List disputes (buyer/seller sees own; admin sees all if used here)",
                "operationId": "91ec5b0ce157f2fe596e9ad05c2d38aa",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (Laravel paginator)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page. Max is capped by server.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/disputes/{id}": {
            "get": {
                "tags": [
                    "Disputes"
                ],
                "summary": "Dispute detail (buyer/seller/admin)",
                "operationId": "7eccf5436f6249e3a90c6a4de1d1f09f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/disputes/{id}/messages": {
            "post": {
                "tags": [
                    "Disputes"
                ],
                "summary": "Add message to dispute (buyer/seller/admin)",
                "operationId": "ef74c44db7839f1b874aa901cdd08d8a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Saya lampirkan foto barang"
                                    },
                                    "media_url": {
                                        "type": "string",
                                        "example": "/storage/uploads/proof.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/orders/{id}/escrow": {
            "get": {
                "tags": [
                    "Escrow"
                ],
                "summary": "Get escrow ledger & balance for an order",
                "operationId": "e8915ee22da6f07089ebdd091d8c497f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/orders/{id}/escrow/hold": {
            "post": {
                "tags": [
                    "Escrow"
                ],
                "summary": "Hold funds (simulate payment success)",
                "operationId": "b2f89198dd1129f8467da06f6ee68db4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/orders/{id}/escrow/release": {
            "post": {
                "tags": [
                    "Escrow"
                ],
                "summary": "Release funds to seller (admin)",
                "operationId": "49e68b9bab1001a3235e2efd78eb9537",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/orders/{id}/escrow/refund": {
            "post": {
                "tags": [
                    "Escrow"
                ],
                "summary": "Refund funds to buyer (admin)",
                "operationId": "8c3e52fad0d7c542f508f0445507abf6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Gunakan Bearer token dari endpoint login/register",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}